+++ scrubadub/detectors/tagged.py
@@ -1,3 +1,4 @@
+import re
 import sys
 import copy
 
@@ -9,35 +10,37 @@ else:
     from typing_extensions import TypedDict
 
 from .base import Detector
-from ..filth.known import KnownFilth
+from ..filth.base import Filth
+from ..filth.tagged import TaggedEvaluationFilth
 
 KnownFilthItem = TypedDict(
     'KnownFilthItem',
     {
         'match': str,
-        'match_end': Optional[str],
-        'limit': Optional[int],
-        'ignore_case': Optional[bool],
+        'match_end': str,
+        'limit': int,
+        'ignore_case': bool,
+        'ignore_whitespace': bool,
+        'ignore_partial_word_matches': bool,
         'filth_type': str
     },
     total=False,
 )
 
 
-class KnownFilthDetector(Detector):
-    """Use this ``Detector`` to find some known filth in the text.
+class TaggedEvaluationFilthDetector(Detector):
+    """Use this ``Detector`` to find tag filth as true ``Filth``. This is useful when you want evaluate the
+    effectiveness of a Detector using Filth that has been selected by a human.
 
-    This is useful in two situations:
-
-        1. If you have known Filth that you want to remove from text (Such as a list of employee numbers)
-        2. If you want evaluate the effectiveness of a Detector using Filth selected by a human.
+    Results from this detector are used as the "truth" against which the other detectos are compared. This is done in
+    ``scrubadub.comparison.get_filth_classification_report`` where the detecton accuracies are calculated.
 
     An example of how to use this detector is given below:
 
     >>> import scrubadub, scrubadub.comparison, scrubadub.detectors.text_blob
     >>> scrubber = scrubadub.Scrubber(detector_list=[
     ...     scrubadub.detectors.TextBlobNameDetector(name='name_detector'),
-    ...     scrubadub.detectors.KnownFilthDetector([
+    ...     scrubadub.detectors.TaggedEvaluationFilthDetector([
     ...         {'match': 'Tom', 'filth_type': 'name'},
     ...         {'match': 'tom@example.com', 'filth_type': 'email'},
     ...     ]),
@@ -58,10 +61,10 @@ class KnownFilthDetector(Detector):
     inialising a ``Scrubber``.
     """
 
-    filth_cls = KnownFilth
-    name = 'known'
+    filth_cls = TaggedEvaluationFilth
+    name = 'tagged'
 
-    def __init__(self, known_filth_items: Optional[List[KnownFilthItem]] = None, **kwargs):
+    def __init__(self, known_filth_items: List[KnownFilthItem], **kwargs):
         """Initialise the ``Detector``.
 
         :param known_filth_items: A list of dictionaries that describe items to be searched for in the dirty text. The
@@ -73,7 +76,10 @@ class KnownFilthDetector(Detector):
             ``{'match': 'aaa', 'filth_type': 'name'}`` will search for an exact match to aaa and return it as a
             ``NameFilth``, where as ``{'match': 'aaa', 'match_end': 'zzz', 'filth_type': 'name'}`` will search for
             `aaa` followed by up to 150 characters followed by `zzz`, which would match both `aaabbbzzz` and `aaazzz`.
-        :type known_filth_items: list of dicts, optional
+        :type known_filth_items: list of dicts
+        :param tagged_filth: Whether the filth has been tagged and should be used as truth when calculating filth
+            finding accuracies.
+        :type tagged_filth: bool, default True
         :param name: Overrides the default name of the :class:``Detector``
         :type name: str, optional
         :param locale: The locale of the documents in the format: 2 letter lower-case language code followed by an
@@ -81,24 +87,60 @@ class KnownFilthDetector(Detector):
         :type locale: str, optional
         """
         super().__init__(**kwargs)
-        if known_filth_items is None:
-            known_filth_items = []
 
         for item in known_filth_items:
+            ignore_whitespace = 'ignore_whitespace' in item and item['ignore_whitespace'] is True
             if 'match' not in item or 'filth_type' not in item:
                 raise KeyError("Each known filth item (dict) needs both keys 'match' and 'filth_type'.")
             if not isinstance(item['match'], str):
                 raise ValueError("The value of 'match' in each KnownItem should be a string. "
                                  "Current value: " + item['match'].__repr__())
+            if not isinstance(item['filth_type'], str):
+                raise ValueError("The value of 'filth_type' in each KnownItem should be a string. "
+                                 "Current value: " + item['filth_type'].__repr__())
+            if ignore_whitespace:
+                item['match'] = " ".join(item['match'].split())
+            else:
+                item['match'] = item['match'].strip()
+            item['filth_type'] = item['filth_type'].strip()
             if 'match_end' in item:
                 if not isinstance(item['match_end'], str):
                     raise ValueError("The value of 'match_end' in each KnownItem should be a string. "
                                      "Current value: " + item['match_end'].__repr__())
+                if ignore_whitespace:
+                    item['match_end'] = " ".join(item['match_end'].split())
+
             for key in item.keys():
-                if key not in ['match', 'match_end', 'limit', 'filth_type', 'ignore_case']:
+                if key not in ['match', 'match_end', 'limit', 'filth_type', 'ignore_case', 'ignore_whitespace',
+                               'ignore_partial_word_matches']:
                     raise KeyError("Unexpected key '{}' in the known filth item.".format(key))
 
-        self._known_filth_items = known_filth_items
+        self._known_filth_items = self.dedup_dicts(known_filth_items)
+
+    @staticmethod
+    def dedup_dicts(known_filth_items: List[KnownFilthItem]) -> List[KnownFilthItem]:
+        # It would be nicer to do this with a set, but sets and dictionaries dont work well together, plus this way
+        # we get to keep the typing info associated to these dicts.
+        deduped = []  # type: List[KnownFilthItem]
+        for item in known_filth_items:
+            if item not in deduped:
+                deduped.append(item)
+
+        return deduped
+
+    def create_filth(
+            self, start_location: int, end_location: int, text: str, comparison_type: Optional[str],
+            detector_name: str, document_name: Optional[str], locale: str
+    ) -> Filth:
+        return TaggedEvaluationFilth(
+            start_location,
+            end_location,
+            text,
+            comparison_type=comparison_type,
+            detector_name=detector_name,
+            document_name=document_name,
+            locale=locale,
+        )
 
     def _find_all(
             self,
@@ -107,7 +149,9 @@ class KnownFilthDetector(Detector):
             comparison_type: Optional[str] = None,
             document_name: Optional[str] = None,
             ignore_case: bool = False,
-    ) -> Generator[KnownFilth, None, None]:
+            ignore_whitespace: bool = False,
+            ignore_partial_word_matches: bool = False,
+    ) -> Generator[Filth, None, None]:
         """Yield filth for each match to substr in text."""
 
         text_orig = copy.copy(text)
@@ -115,22 +159,25 @@ class KnownFilthDetector(Detector):
             text = text.lower()
             substr = substr.lower()
 
-        substr_len = len(substr)
-        start_location = text.find(substr)
+        if ignore_whitespace:
+            # We change any white space in the original with "\s+" that has to match one or more whitespace chars
+            substr = '\\s+'.join([re.escape(token) for token in substr.split()])
+        else:
+            substr = re.escape(substr)
 
-        while start_location >= 0:
-            yield KnownFilth(
-                start_location,
-                start_location + substr_len,
-                text_orig[start_location:start_location + substr_len],
+        if ignore_partial_word_matches:
+            substr = f"\\b{substr}\\b"
+
+        matches = re.finditer(substr, text)
+        for match in matches:
+            yield self.create_filth(
+                match.span()[0],
+                match.span()[1],
+                text_orig[match.span()[0]:match.span()[1]],
                 comparison_type=comparison_type,
                 detector_name=self.name,
                 document_name=document_name,
                 locale=self.locale,
-            )
-            start_location = text.find(
-                substr,
-                start_location + substr_len
             )
 
     def _find_all_between(
@@ -142,7 +189,9 @@ class KnownFilthDetector(Detector):
             comparison_type: Optional[str] = None,
             document_name: Optional[str] = None,
             ignore_case: bool = False,
-    ) -> Generator[KnownFilth, None, None]:
+            ignore_whitespace: bool = False,
+            ignore_partial_word_matches: bool = False,
+    ) -> Generator[Filth, None, None]:
         """Yield filth for text between (and including)
         substr_start and substr_end, but only if the text
         between the two is less than limit characters.
@@ -153,37 +202,35 @@ class KnownFilthDetector(Detector):
             substr_start = substr_start.lower()
             substr_end = substr_end.lower()
 
-        substr_start_len = len(substr_start)
-        substr_end_len = len(substr_end)
-        start_location = text.find(substr_start)
+        if ignore_whitespace:
+            # We change any white space in the original with "\s+" that has to match one or more whitespace chars
+            substr_start = '\\s+'.join([re.escape(token) for token in substr_start.split()])
+            substr_end = '\\s+'.join([re.escape(token) for token in substr_end.split()])
+        else:
+            substr_start = re.escape(substr_start)
+            substr_end = re.escape(substr_end)
 
-        while start_location >= 0:
-            end_location = text.find(
-                substr_end,
-                start_location + substr_start_len,
-                start_location + substr_start_len + limit + substr_end_len
+        if ignore_partial_word_matches:
+            substr_start = f"\\b{substr_start}\\b"
+            substr_end = f"\\b{substr_end}\\b"
+
+        matches = re.finditer(f"({substr_start})(.{{0,{limit}}})({substr_end})", text)
+        for match in matches:
+            yield self.create_filth(
+                match.span()[0],
+                match.span()[1],
+                text_orig[match.span()[0]:match.span()[1]],
+                comparison_type=comparison_type,
+                detector_name=self.name,
+                document_name=document_name,
+                locale=self.locale,
             )
-            if end_location >= 0:
-                yield KnownFilth(
-                    start_location,
-                    end_location + substr_end_len,
-                    text_orig[start_location:end_location + substr_end_len],
-                    comparison_type=comparison_type,
-                    detector_name=self.name,
-                    document_name=document_name,
-                    locale=self.locale,
-                )
-                next_search_start = end_location + substr_end_len
-            else:
-                next_search_start = start_location + substr_start_len
-
-            start_location = text.find(substr_start, next_search_start)
 
     def iter_filth(
             self,
             text: str,
             document_name: Optional[str] = None
-    ) -> Generator[KnownFilth, None, None]:
+    ) -> Generator[Filth, None, None]:
         """Yields discovered filth in the provided ``text``.
 
         :param text: The dirty text to clean.
@@ -195,8 +242,9 @@ class KnownFilthDetector(Detector):
         """
         for pii_item in self._known_filth_items:
             # could also implement other types in here too
-            ignore_case = pii_item.get('ignore_case', None)
-            ignore_case = ignore_case if ignore_case is not None else False
+            ignore_case = pii_item.get('ignore_case', False)
+            ignore_whitespace = pii_item.get('ignore_whitespace', False)
+            ignore_partial_word_matches = pii_item.get('ignore_partial_word_matches', False)
             if 'match' in pii_item and 'match_end' in pii_item and pii_item['match_end'] is not None \
                     and len(pii_item['match_end']) > 0:
                 yield from self._find_all_between(
@@ -207,6 +255,8 @@ class KnownFilthDetector(Detector):
                         comparison_type=pii_item.get('filth_type', None),
                         document_name=document_name,
                         ignore_case=ignore_case,
+                        ignore_whitespace=ignore_whitespace,
+                        ignore_partial_word_matches=ignore_partial_word_matches,
                 )
             elif 'match' in pii_item:
                 yield from self._find_all(
@@ -215,6 +265,8 @@ class KnownFilthDetector(Detector):
                         comparison_type=pii_item.get('filth_type', None),
                         document_name=document_name,
                         ignore_case=ignore_case,
+                        ignore_whitespace=ignore_whitespace,
+                        ignore_partial_word_matches=ignore_partial_word_matches,
                 )
             else:
                 raise ValueError(